Typography: Transform plaintext '---' into em-dash
And '--' into en-dash. It's easy enough for me to type a literal
em- or en-dash, but I used to write markdown using triple or double
hyphens instead because its looks nicer in plaintext and Pandoc has
an option to make this same transformation---but I no longer use
Pandoc for this blog. Reimplementing the adjustment seemed less
hassle than changing the markdown files themselves.